Use backported -Wconf flag
Scala 2.13 introduced a mechanism to provide fine-grained control over
what warnings are checked for during compilation and how to handle them
when they do appear.  This was later backported to 2.12.13

We want to be able to have warnings be fatal except for deprecation
warnings since the warnings come in part for external code, the
deprecated APIs are still currently usable, and adapting our code to
account for deprecated APIs may be non-trivial.  We had some code in SBT
which tried to do the same by monitoring output, but that acheived only
mixed success.

More details on -Wconf described in [Configuring and Suppressing Warnings](https://www.scala-lang.org/2021/01/12/configuring-and-suppressing-warnings.html)